wip refactor(jest): migrate from js to ts
