+++ .eslintrc.yml
@@ -2,7 +2,7 @@ parser: '@typescript-eslint/parser'
 parserOptions:
   ecmaVersion: 2018
   sourceType: module
-  project: tsconfig.json
+  project: tsconfig.eslint.json
 
 env:
   es6: true
@@ -13,8 +13,7 @@ extends:
   - eslint:recommended
   # - plugin:@typescript-eslint/recommended
   # - plugin:@typescript-eslint/recommended-requiring-type-checking
-  # - plugin:jest/recommended
-  # - prettier/@typescript-eslint
+  - prettier/@typescript-eslint
   - prettier # must go last, to turn off some previous rules
 
 rules:
@@ -23,5 +22,6 @@ rules:
   no-use-before-define: 0
   require-await: error
   '@typescript-eslint/no-use-before-define': 0
-# TODO
-# Migrate chai, jest
+
+ignorePatterns:
+  - dist
