--- packages/jest-openapi/.eslintignore
@@ -1,1 +1,0 @@
-reports
