--- packages/jest-openapi/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./src');
