+++ packages/jest-openapi/index.ts
@@ -1,0 +1,3 @@
+import src from './src';
+
+export default src;
