+++ packages/jest-openapi/jest.config.ts
@@ -1,4 +1,6 @@
-module.exports = {
+import type { Config } from '@jest/types';
+
+const config: Config.InitialOptions = {
   preset: 'ts-jest',
   testEnvironment: 'node',
   rootDir: '..',
@@ -16,3 +18,5 @@ module.exports = {
     },
   },
 };
+
+export default config;
