+++ packages/jest-openapi/package.json
@@ -7,15 +7,16 @@
   "scripts": {
     "clean": "rimraf dist",
     "format": "prettier --write ../../ --ignore-path ../../.prettierignore",
-    "lint": "tsc --noEmit && eslint .",
+    "lint": "tsc --noEmit --project tsconfig.eslint.json && eslint .",
     "lint:fix": "yarn lint --fix",
     "build": "tsc",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
+    "test:coverage": "yarn clean && jest --coverage",
     "test:coverage:browse": "yarn test:coverage; open ../coverage/lcov-report/index.html",
     "test:precommit": "yarn format && yarn lint && yarn test:coverage",
-    "test:ci": "yarn test:precommit"
+    "test:ci": "yarn test:precommit",
+    "prepack": "yarn build"
   },
   "repository": "https://github.com/openapi-library/OpenAPIValidators/tree/master/packages/jest-openapi",
   "author": "rwalle61 <richard.lh.waller@gmail.com>",
@@ -36,7 +37,7 @@
   },
   "homepage": "https://github.com/openapi-library/OpenAPIValidators/tree/master/packages/jest-openapi#readme",
   "files": [
-    "dist/**/*"
+    "dist"
   ],
   "devDependencies": {
     "@types/jest": "^26.0.20",
@@ -51,6 +52,7 @@
     "request-promise": "^4.2.6",
     "supertest": "^6.0.0",
     "ts-jest": "^26.5.3",
+    "ts-node": "^9.1.1",
     "typescript": "^4.2.3"
   },
   "dependencies": {
