--- packages/jest-openapi/src/utils.js
@@ -1,10 +1,0 @@
-const { inspect } = require('util');
-
-const stringify = (obj) => inspect(obj, { showHidden: false, depth: null });
-
-const joinWithNewLines = (...lines) => lines.join('\n\n');
-
-module.exports = {
-  stringify,
-  joinWithNewLines,
-};
